Move _Context back to song_loader module.
